Simplified the audio check
Added a useEffect and moved the pronunciation and buttonDisabled variables to state. I was having an issue where the phonetics array was 0 and it would skip the map entirely thus leaving the button enabled, refactoring this help me get more familiar with the .find method to avoid the map completely.